operator: integrate openapi server implementation with cilium-operator
