+++ Documentation/cmdref/cilium-operator-aws_metrics.md
@@ -1,0 +1,17 @@
+<!-- This file was autogenerated via cilium-operator --cmdref, do not edit manually-->
+
+## cilium-operator-aws metrics
+
+Access metric status of the operator
+
+### Options
+
+```
+  -h, --help   help for metrics
+```
+
+### SEE ALSO
+
+* [cilium-operator-aws](cilium-operator-aws.html)	 - Run cilium-operator-aws
+* [cilium-operator-aws metrics list](cilium-operator-aws_metrics_list.html)	 - List all metrics for the operator
+
