+++ Documentation/cmdref/cilium-operator-azure_metrics.md
@@ -1,0 +1,17 @@
+<!-- This file was autogenerated via cilium-operator --cmdref, do not edit manually-->
+
+## cilium-operator-azure metrics
+
+Access metric status of the operator
+
+### Options
+
+```
+  -h, --help   help for metrics
+```
+
+### SEE ALSO
+
+* [cilium-operator-azure](cilium-operator-azure.html)	 - Run cilium-operator-azure
+* [cilium-operator-azure metrics list](cilium-operator-azure_metrics_list.html)	 - List all metrics for the operator
+
