+++ Documentation/cmdref/cilium-operator-azure_metrics_list.md
@@ -1,0 +1,22 @@
+<!-- This file was autogenerated via cilium-operator --cmdref, do not edit manually-->
+
+## cilium-operator-azure metrics list
+
+List all metrics for the operator
+
+```
+cilium-operator-azure metrics list [flags]
+```
+
+### Options
+
+```
+  -h, --help                   help for list
+  -p, --match-pattern string   Show only metrics whose names match matchpattern
+  -o, --output string          json| jsonpath='{}'
+```
+
+### SEE ALSO
+
+* [cilium-operator-azure metrics](cilium-operator-azure_metrics.html)	 - Access metric status of the operator
+
