+++ Documentation/cmdref/cilium-operator-generic.md
@@ -65,3 +65,7 @@ cilium-operator-generic [flags]
       --version                                   Print version information
 ```
 
+### SEE ALSO
+
+* [cilium-operator-generic metrics](cilium-operator-generic_metrics.html)	 - Access metric status of the operator
+
