+++ Documentation/cmdref/cilium-operator-generic_metrics.md
@@ -1,0 +1,17 @@
+<!-- This file was autogenerated via cilium-operator --cmdref, do not edit manually-->
+
+## cilium-operator-generic metrics
+
+Access metric status of the operator
+
+### Options
+
+```
+  -h, --help   help for metrics
+```
+
+### SEE ALSO
+
+* [cilium-operator-generic](cilium-operator-generic.html)	 - Run cilium-operator-generic
+* [cilium-operator-generic metrics list](cilium-operator-generic_metrics_list.html)	 - List all metrics for the operator
+
