+++ Documentation/cmdref/cilium-operator_metrics.md
@@ -1,0 +1,17 @@
+<!-- This file was autogenerated via cilium-operator --cmdref, do not edit manually-->
+
+## cilium-operator metrics
+
+Access metric status of the operator
+
+### Options
+
+```
+  -h, --help   help for metrics
+```
+
+### SEE ALSO
+
+* [cilium-operator](cilium-operator.html)	 - Run cilium-operator
+* [cilium-operator metrics list](cilium-operator_metrics_list.html)	 - List all metrics for the operator
+
