+++ Documentation/cmdref/cilium-operator_metrics_list.md
@@ -1,0 +1,22 @@
+<!-- This file was autogenerated via cilium-operator --cmdref, do not edit manually-->
+
+## cilium-operator metrics list
+
+List all metrics for the operator
+
+```
+cilium-operator metrics list [flags]
+```
+
+### Options
+
+```
+  -h, --help                   help for list
+  -p, --match-pattern string   Show only metrics whose names match matchpattern
+  -o, --output string          json| jsonpath='{}'
+```
+
+### SEE ALSO
+
+* [cilium-operator metrics](cilium-operator_metrics.html)	 - Access metric status of the operator
+
